Add method in LessonRepository create validation for createLesson and updateLesson

+++ CharlieBackend.Business/Services/LessonService.cs
@@ -9,6 +9,7 @@ using CharlieBackend.Business.Services.Interfaces;
 using CharlieBackend.Data.Repositories.Impl.Interfaces;
 using CharlieBackend.Core.DTO.Lesson;
 using CharlieBackend.Core.Models.ResultModel;
+using CharlieBackend.Core.DTO.Visit;
 
 namespace CharlieBackend.Business.Services
 {
@@ -44,18 +45,29 @@ namespace CharlieBackend.Business.Services
               
                 if (foundMentor == null)
                 {
-                    throw new Exception("Mentor is not found"); // return Result<T>
+                    return Result<LessonDto>.GetError(ErrorCode.NotFound ,"Mentor is not found"); // return Result<T>
                 }
                 
                 var foundStudentGroup = await _unitOfWork.StudentGroupRepository.GetByIdAsync(lessonDto.StudentGroupId);
                 
                 if (foundStudentGroup == null)
                 {
-                    throw new Exception("Student group is not found"); // 
+                    return Result<LessonDto>.GetError(ErrorCode.NotFound, "Student group is not found");
                 }
+               
                 if (foundStudentGroup.StudentsOfStudentGroups.Count != lessonDto.LessonVisits.Count)
                 {
-                    throw new Exception("something wromg with a Lesso Visits");
+                    return Result<LessonDto>.GetError(ErrorCode.ValidationError, "Something wrong with a Lesson Visits");
+                }
+
+                if (!await CheckStudentsID(lessonDto.LessonVisits, lessonDto.StudentGroupId))
+                {
+                    return Result<LessonDto>.GetError(ErrorCode.NotFound, "Someone of student is not found");
+                }
+
+                if (createdLessonEntity.LessonDate < DateTime.Now)
+                {
+                    return Result<LessonDto>.GetError(ErrorCode.ValidationError, "Lesson date is incorrect");
                 }
 
                 _unitOfWork.LessonRepository.Add(createdLessonEntity);
@@ -74,11 +86,11 @@ namespace CharlieBackend.Business.Services
 
                 return Result<LessonDto>.GetSuccess(_mapper.Map<LessonDto>(createdLessonEntity));
             }
-            catch(Exception ex)
+            catch
             {
                 _unitOfWork.Rollback();
 
-                return Result<LessonDto>.GetError(ErrorCode., ex.Message);
+                return null;
             }
         }
 
@@ -138,12 +150,12 @@ namespace CharlieBackend.Business.Services
                     }
                 }
 
-                if (lessonModel.LessonDate != default(DateTime))
+                if (lessonModel.LessonDate != default(DateTime) && lessonModel.LessonDate < DateTime.Now)
                 {
                     foundLesson.LessonDate = lessonModel.LessonDate;
                 }
 
-                if (lessonModel.LessonVisits != null)
+                if (lessonModel.LessonVisits != null && !await CheckStudentsID(lessonModel.LessonVisits, (long)foundLesson.StudentGroupId))
                 {
                     await _unitOfWork.VisitRepository.DeleteWhereLessonIdAsync(foundLesson.Id);
 
@@ -186,5 +198,31 @@ namespace CharlieBackend.Business.Services
             return Result<IList<StudentLessonDto>>.GetSuccess( _mapper.Map<IList<StudentLessonDto>>(studentLessonModels));
         }
 
+        private async Task<bool> CheckStudentsID(IList<VisitDto> visit, long groupId)
+        {
+            List<long> studentsId = new List<long>();
+            
+            foreach (var item in visit)
+            {
+                studentsId.Add((long)item.StudentId);
+            }
+
+            var checkStudentInGroup = await _unitOfWork.StudentGroupRepository.GetAllStudentInGroup(groupId);
+
+            if (checkStudentInGroup.Count != studentsId.Count)
+            {
+                return false;
+            }
+
+            foreach (var item in studentsId)
+            {
+                if (!checkStudentInGroup.Contains(item))
+                {
+                    return false;
+                }
+            }            
+
+            return true;
+        }
     }
 }
